To add error handling for curriculum items

+++ src/school-years/school-years.service.ts
@@ -1,10 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateSchoolYearDto } from './dto/create-school-year.dto';
 import { UpdateSchoolYearDto } from './dto/update-school-year.dto';
+import { ErrorCode } from 'src/common/error-codes';
+import { ApiErrorResponseBuilder } from 'src/common/api-error-builder';
 
 @Injectable()
 export class SchoolYearsService {
+  private readonly logger = new Logger(SchoolYearsService.name);
   constructor(private prisma: PrismaService) {}
 
   async create(createSchoolYearDto: CreateSchoolYearDto) {
@@ -69,12 +72,26 @@ export class SchoolYearsService {
   }
 
   async remove(id: string) {
+
+    // check if the school year is used in course sections
+    const courseSections = await this.prisma.courseSection.findMany({
+      where: { schoolYearId: id },
+    });
+    if (courseSections.length > 0) {
+      const errorResponse = ApiErrorResponseBuilder.create(
+        ErrorCode.SCYA,
+        'School year is used in course sections'
+      )
+        .withLogger(this.logger)
+        .build();
+      throw new BadRequestException(errorResponse);
+    }
     try {
       return await this.prisma.schoolYear.delete({
         where: { id },
-        include: {
+        /* include: {
           terms: true,
-        },
+        }, */
       });
     } catch (error) {
       throw new NotFoundException(`School year with ID ${id} not found`);
