+++ src/time-blocks/time-blocks.service.ts
@@ -1,10 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, ConflictException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateTimeBlockDto } from './dto/create-time-block.dto';
 import { UpdateTimeBlockDto } from './dto/update-time-block.dto';
+import { ApiErrorResponseBuilder } from 'src/common/api-error-builder';
+import { ErrorCode } from 'src/common/error-codes';
 
 @Injectable()
 export class TimeBlocksService {
+  private readonly logger = new Logger(TimeBlocksService.name);
   constructor(private prisma: PrismaService) {}
 
   async create(createTimeBlockDto: CreateTimeBlockDto) {
@@ -27,7 +30,13 @@ export class TimeBlocksService {
     });
 
     if (!timeBlock) {
-      throw new NotFoundException(`Time block with ID ${id} not found`);
+      const errorResponse = ApiErrorResponseBuilder.create(
+        ErrorCode.TBDB,
+        `Time block with ID ${id} not found`
+      )
+        .withLogger(this.logger)
+        .build();
+      throw new NotFoundException(errorResponse);
     }
 
     return timeBlock;
@@ -45,12 +54,49 @@ export class TimeBlocksService {
   }
 
   async remove(id: string) {
-    try {
+ 
+      // check if the time block is active
+      const timeBlock = await this.prisma.timeBlock.findUnique({
+        where: { id },
+      });
+
+      if (!timeBlock) {
+        const errorResponse = ApiErrorResponseBuilder.create(
+          ErrorCode.TBDB,
+          `Time block with ID ${id} not found`
+        )
+          .withLogger(this.logger)
+          .build();
+        throw new NotFoundException(errorResponse);
+      }
+
+      // check if the time block is used in course sections 
+      const courseSections = await this.prisma.courseSection.findMany({
+        where: { timeBlockId: id },
+      });
+      
+      if (courseSections.length > 0) {
+        const errorResponse = ApiErrorResponseBuilder.create(
+          ErrorCode.TBDA,
+          'Cannot delete time block that is used in course sections'
+        )
+          .withLogger(this.logger)
+          .build();
+        throw new ConflictException(errorResponse);
+      }
+      try {
       return await this.prisma.timeBlock.delete({
         where: { id },
       });
     } catch (error) {
-      throw new NotFoundException(`Time block with ID ${id} not found`);
+      const errorResponse = ApiErrorResponseBuilder.create(
+        ErrorCode.TBDB,
+       `Time block with ID ${id} not found`
+      )
+        .withLogger(this.logger)
+        .build();
+      
+      throw new NotFoundException(errorResponse);
     }
   }
